refactor: Extract type inheritance utilities into their own module
